requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-appconf==1.0.5
django-crispy-forms==2.0
django-mathfilters==1.0.0
django-select2==8.1.1
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
pycodestyle==2.10.0
python-barcode==0.14.0
python-decouple==3.8
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
tomli==2.0.1
urllib3==1.26.15
whitenoise==6.4.0
